updated harris

+++ soda_codes/camera_mini_opt/our_code/camera_mini_opt_compute_units.h
@@ -4,59 +4,59 @@
 
 // Generated compute units...
 hw_uint<16> raw_generated_compute(hw_uint<16*1> raw_oc) {
-  return raw_oc.get<16, 0>();
+  return uint16_t(uint16_t(raw_oc.get<16, 0>()));
 }
 
 hw_uint<16> blue_generated_compute(hw_uint<16*9> raw) {
-  return ((((((((raw.get<16, 0>() + raw.get<16, 1>()) + raw.get<16, 2>()) + raw.get<16, 3>()) + raw.get<16, 4>()) + raw.get<16, 5>()) + raw.get<16, 6>()) + raw.get<16, 7>()) + raw.get<16, 8>());
+  return uint16_t(((((((((uint16_t(raw.get<16, 0>()) + uint16_t(raw.get<16, 1>())) + uint16_t(raw.get<16, 2>())) + uint16_t(raw.get<16, 3>())) + uint16_t(raw.get<16, 4>())) + uint16_t(raw.get<16, 5>())) + uint16_t(raw.get<16, 6>())) + uint16_t(raw.get<16, 7>())) + uint16_t(raw.get<16, 8>())));
 }
 
 hw_uint<16> red_generated_compute(hw_uint<16*9> raw) {
-  return ((((((((raw.get<16, 0>() + raw.get<16, 1>()) + raw.get<16, 2>()) + raw.get<16, 3>()) + raw.get<16, 4>()) + raw.get<16, 5>()) + raw.get<16, 6>()) + raw.get<16, 7>()) + raw.get<16, 8>());
+  return uint16_t(((((((((uint16_t(raw.get<16, 0>()) + uint16_t(raw.get<16, 1>())) + uint16_t(raw.get<16, 2>())) + uint16_t(raw.get<16, 3>())) + uint16_t(raw.get<16, 4>())) + uint16_t(raw.get<16, 5>())) + uint16_t(raw.get<16, 6>())) + uint16_t(raw.get<16, 7>())) + uint16_t(raw.get<16, 8>())));
 }
 
 hw_uint<16> green_generated_compute(hw_uint<16*9> raw) {
-  return ((((((((raw.get<16, 0>() + raw.get<16, 1>()) + raw.get<16, 2>()) + raw.get<16, 3>()) + raw.get<16, 4>()) + raw.get<16, 5>()) + raw.get<16, 6>()) + raw.get<16, 7>()) + raw.get<16, 8>());
+  return uint16_t(((((((((uint16_t(raw.get<16, 0>()) + uint16_t(raw.get<16, 1>())) + uint16_t(raw.get<16, 2>())) + uint16_t(raw.get<16, 3>())) + uint16_t(raw.get<16, 4>())) + uint16_t(raw.get<16, 5>())) + uint16_t(raw.get<16, 6>())) + uint16_t(raw.get<16, 7>())) + uint16_t(raw.get<16, 8>())));
 }
 
 hw_uint<16> blue_bx_generated_compute(hw_uint<16*3> blue) {
-  return ((blue.get<16, 0>() + blue.get<16, 1>()) + blue.get<16, 2>());
+  return uint16_t(((uint16_t(blue.get<16, 0>()) + uint16_t(blue.get<16, 1>())) + uint16_t(blue.get<16, 2>())));
 }
 
 hw_uint<16> red_bx_generated_compute(hw_uint<16*3> red) {
-  return ((red.get<16, 0>() + red.get<16, 1>()) + red.get<16, 2>());
+  return uint16_t(((uint16_t(red.get<16, 0>()) + uint16_t(red.get<16, 1>())) + uint16_t(red.get<16, 2>())));
 }
 
 hw_uint<16> green_bx_generated_compute(hw_uint<16*3> green) {
-  return ((green.get<16, 0>() + green.get<16, 1>()) + green.get<16, 2>());
+  return uint16_t(((uint16_t(green.get<16, 0>()) + uint16_t(green.get<16, 1>())) + uint16_t(green.get<16, 2>())));
 }
 
 hw_uint<16> blue_by_generated_compute(hw_uint<16*3> blue_bx) {
-  return ((blue_bx.get<16, 0>() + blue_bx.get<16, 1>()) + blue_bx.get<16, 2>());
+  return uint16_t(((uint16_t(blue_bx.get<16, 0>()) + uint16_t(blue_bx.get<16, 1>())) + uint16_t(blue_bx.get<16, 2>())));
 }
 
 hw_uint<16> red_by_generated_compute(hw_uint<16*3> red_bx) {
-  return ((red_bx.get<16, 0>() + red_bx.get<16, 1>()) + red_bx.get<16, 2>());
+  return uint16_t(((uint16_t(red_bx.get<16, 0>()) + uint16_t(red_bx.get<16, 1>())) + uint16_t(red_bx.get<16, 2>())));
 }
 
 hw_uint<16> green_by_generated_compute(hw_uint<16*3> green_bx) {
-  return ((green_bx.get<16, 0>() + green_bx.get<16, 1>()) + green_bx.get<16, 2>());
+  return uint16_t(((uint16_t(green_bx.get<16, 0>()) + uint16_t(green_bx.get<16, 1>())) + uint16_t(green_bx.get<16, 2>())));
 }
 
 hw_uint<16> blue_diff_generated_compute(hw_uint<16*1> blue, hw_uint<16*1> blue_by) {
-  return (blue_by.get<16, 0>() - blue.get<16, 0>());
+  return uint16_t((uint16_t(blue_by.get<16, 0>()) - uint16_t(blue.get<16, 0>())));
 }
 
 hw_uint<16> red_diff_generated_compute(hw_uint<16*1> red, hw_uint<16*1> red_by) {
-  return (red_by.get<16, 0>() - red.get<16, 0>());
+  return uint16_t((uint16_t(red_by.get<16, 0>()) - uint16_t(red.get<16, 0>())));
 }
 
 hw_uint<16> green_diff_generated_compute(hw_uint<16*1> green, hw_uint<16*1> green_by) {
-  return (green_by.get<16, 0>() - green.get<16, 0>());
+  return uint16_t((uint16_t(green_by.get<16, 0>()) - uint16_t(green.get<16, 0>())));
 }
 
 hw_uint<16> camera_mini_generated_compute(hw_uint<16*1> blue_diff, hw_uint<16*1> green_diff, hw_uint<16*1> red_diff) {
-  return ((red_diff.get<16, 0>() + green_diff.get<16, 0>()) + blue_diff.get<16, 0>());
+  return uint16_t(((uint16_t(red_diff.get<16, 0>()) + uint16_t(green_diff.get<16, 0>())) + uint16_t(blue_diff.get<16, 0>())));
 }
